intercom: refactor into intercom_srv and intercom_client
